Add type-safe Interval to represent the timespan and rolling-average to get bitcoin prices

+++ charts/data/src/main/java/bitcoin/charts/data/network/MarketPriceResponse.kt
@@ -14,5 +14,5 @@ class MarketPriceResponse(
  */
 data class RawPrice(
     @SerializedName("x") val timestamp: Long,
-    @SerializedName("y") val price: Double
+    @SerializedName("y") val price: Float
 )
